backend/requirements.txt
django==5.0.0
djangorestframework==3.14.0
django-cors-headers==4.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
google-api-python-client==2.170.0
google-auth==2.40.3
